Refactor the variable names
